refactor(bifrost): extract inbound/ package layer from app.py
Move route handlers, quota enforcement, and SSE tracking helpers out of
the monolithic app.py into a dedicated inbound/ package:

- inbound/tracking.py — _extract_usage_from_sse_line, _log_request,
  _stream_with_tracking (SSE token-tracking helpers)
- inbound/routes.py   — _check_quotas, _check_model_access, create_router()
  (all route handlers via APIRouter factory)
- app.py              — now a thin factory: _build_usage_store,
  _pricing_overrides, create_app (wires inbound to FastAPI)

Update test imports to reference the new module locations.

Closes NIU-513.
