+++ src/bifrost/inbound/__init__.py
@@ -1,1 +1,7 @@
-"""Bifröst inbound interface modules."""
+"""Bifröst inbound interface modules.
+
+Sub-modules:
+  chat_completions — OpenAI Chat Completions request/response translation
+  routes           — FastAPI route handlers and quota/access enforcement
+  tracking         — SSE token-tracking helpers
+"""
